move list of bots into a separate file
