+++ utopian-bot.js
@@ -237,173 +237,10 @@ conn.once('open', function ()
                                   Post
                                     .list({ skip: 0, limit: contributionsCount, query })
                                     .then(contributions => {
-                                      const bots = [
-                                        'analisa',
-                                        'animus',
-                                        'appreciator',
-                                        'arama',
-                                        'ausbitbot',
-                                        'bago',
-                                        'bambam808',
-                                        'banjo',
-                                        'barrie',
-                                        'bellyrub',
-                                        'besttocome215',
-                                        'bierkaart',
-                                        'biskopakon',
-                                        'blackwidow7',
-                                        'blimbossem',
-                                        'boomerang',
-                                        'booster',
-                                        'boostupvote',
-                                        'bowlofbitcoin',
-                                        'bp423',
-                                        'brandybb',
-                                        'brensker',
-                                        'btcvenom',
-                                        'buildawhale',
-                                        'burdok213',
-                                        'businessbot',
-                                        'centerlink',
-                                        'cleverbot',
-                                        'cnbuddy',
-                                        'counterbot',
-                                        'crypto-hangouts',
-                                        'cryptobooty',
-                                        'cryptoholic',
-                                        'cryptoowl',
-                                        'cub1',
-                                        'curationrus',
-                                        'dahrma',
-                                        'davidding',
-                                        'decibel',
-                                        'deutschbot',
-                                        'dirty.hera',
-                                        'discordia',
-                                        'done',
-                                        'drakkald',
-                                        'drotto',
-                                        'earthboundgiygas',
-                                        'edrivegom',
-                                        'emilhoch',
-                                        'eoscrusher',
-                                        'famunger',
-                                        'feedyourminnows',
-                                        'followforupvotes',
-                                        'frontrunner',
-                                        'fuzzyvest',
-                                        'gamerpool',
-                                        'gamerveda',
-                                        'gaming-hangouts',
-                                        'gindor',
-                                        'givemedatsteem',
-                                        'givemesteem1',
-                                        'glitterbooster',
-                                        'gonewhaling',
-                                        'gotvotes',
-                                        'gpgiveaways',
-                                        'gsgaming',
-                                        'guarddog',
-                                        'heelpopulair',
-                                        'helpfulcrypto',
-                                        'idioticbot',
-                                        'ikwindje',
-                                        'ilvacca',
-                                        'inchonbitcoin',
-                                        'ipuffyou',
-                                        'lovejuice',
-                                        'mahabrahma',
-                                        'make-a-whale',
-                                        'makindatsteem',
-                                        'maradaratar',
-                                        'minnowbooster',
-                                        'minnowhelper',
-                                        'minnowpond',
-                                        'minnowpondblue',
-                                        'minnowpondred',
-                                        'minnowsupport',
-                                        'misterwister',
-                                        'moonbot',
-                                        'morwhale',
-                                        'moses153',
-                                        'moyeses',
-                                        'msp-lovebot',
-                                        'msp-shanehug',
-                                        'muxxybot',
-                                        'myday',
-                                        'ninja-whale',
-                                        'ninjawhale',
-                                        'officialfuzzy',
-                                        'perennial',
-                                        'pimpoesala',
-                                        'polsza',
-                                        'portoriko',
-                                        'prambarbara',
-                                        'proctologic',
-                                        'pumpingbitcoin',
-                                        'pushup',
-                                        'qurator',
-                                        'qwasert',
-                                        'raidrunner',
-                                        'ramta',
-                                        'randovote',
-                                        'randowhale',
-                                        'randowhale0',
-                                        'randowhale1',
-                                        'randowhaletrail',
-                                        'randowhaling',
-                                        'reblogger',
-                                        'resteem.bot',
-                                        'resteemable',
-                                        'resteembot',
-                                        'russiann',
-                                        'scamnotifier',
-                                        'scharmebran',
-                                        'siliwilly',
-                                        'sneaky-ninja',
-                                        'sniffo35',
-                                        'soonmusic',
-                                        'spinbot',
-                                        'stackin',
-                                        'steemedia',
-                                        'steemholder',
-                                        'steemit-gamble',
-                                        'steemit-hangouts',
-                                        'steemitgottalent',
-                                        'steemmaker',
-                                        'steemmemes',
-                                        'steemminers',
-                                        'steemode',
-                                        'steemprentice',
-                                        'steemsquad',
-                                        'steemthat',
-                                        'steemvoter',
-                                        'stephen.king989',
-                                        'tabea',
-                                        'tarmaland',
-                                        'timbalabuch',
-                                        'trail1',
-                                        'trail2',
-                                        'trail3',
-                                        'trail4',
-                                        'trail5',
-                                        'trail6',
-                                        'trail7',
-                                        'viraltrend',
-                                        'votey',
-                                        'waardanook',
-                                        'wahyurahadiann',
-                                        'wannabeme',
-                                        'weareone1',
-                                        'whatamidoing',
-                                        'whatupgg',
-                                        'wildoekwind',
-                                        'wiseguyhuh',
-                                        'wistoepon',
-                                        'zdashmash',
-                                        'zdemonz',
-                                        'zhusatriani'
-                                      ];
+
+                                      // get a list of bots
+                                      var list_of_bots = require('./list-of-bots.js');
+                                      var bots = list_of_bots.bots();                                    		
 
                                       const reputation = steem.formatter.reputation(account.reputation);
                                       const categoryStats = categories[post.json_metadata.type];
